KEP-9876: Review cleanup — extract rebroadcast, simplify Skip, tighten visibility
- Extract rebroadcast logic from manager.go into rebroadcast.go for
  easier rebasing against upstream changes.
- Remove reason and logger params from ProtectionSkipTracker.Skip();
  callers now own the V(4) log line with the specific preemption reason.
- Make WithinProtectionWindow private (withinProtectionWindow) since it
  is only called by Skip() within the same package.
