+++ pkg/scheduler/preemption/common/protection.go
@@ -19,10 +19,8 @@ package common
 import (
 	"time"
 
-	"github.com/go-logr/logr"
 	apimeta "k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/klog/v2"
 
 	config "sigs.k8s.io/kueue/apis/config/v1beta2"
 	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta2"
@@ -30,14 +28,14 @@ import (
 	workloadevict "sigs.k8s.io/kueue/pkg/workload/evict"
 )
 
-// WithinProtectionWindow reports whether candidate is protected from preemption
+// withinProtectionWindow reports whether candidate is protected from preemption
 // at time now under the given rule: the PreemptionProtection feature gate is on,
 // the rule is set, the candidate is Admitted and not already Evicted, and its
 // runtime since the Admitted transition is strictly less than minAdmitDuration.
 // A candidate whose runtime is exactly equal to minAdmitDuration is eligible
 // for preemption. Already-evicted candidates are never protected: they are the
 // preferred zero-cost preemption victims.
-func WithinProtectionWindow(candidate *kueue.Workload, rule *config.PreemptionProtectionPolicy, now time.Time) bool {
+func withinProtectionWindow(candidate *kueue.Workload, rule *config.PreemptionProtectionPolicy, now time.Time) bool {
 	if rule == nil || rule.MinAdmitDuration == nil {
 		return false
 	}
@@ -56,7 +54,7 @@ func WithinProtectionWindow(candidate *kueue.Workload, rule *config.PreemptionPr
 
 // ProtectionExpiry returns the time at which the candidate's protection under
 // the given rule expires: the Admitted condition transition time plus the
-// rule's minAdmitDuration. Mirroring WithinProtectionWindow, it returns the
+// rule's minAdmitDuration. Mirroring withinProtectionWindow, it returns the
 // zero time if the rule is unset or the candidate's Admitted condition is not
 // True.
 func ProtectionExpiry(candidate *kueue.Workload, rule *config.PreemptionProtectionPolicy) time.Time {
@@ -81,23 +79,17 @@ type ProtectionSkipTracker struct {
 
 // Skip reports whether candidate is within its protection window under rule
 // at time now. When it is, Skip records the candidate's protection expiry
-// (keeping the earliest across calls) and logs the skip at V(4) with the
-// preemption reason the candidate would otherwise have been preempted with.
-// Callers must evaluate all other preemption criteria first, so that a skip
-// means protection was the sole blocker. A nil tracker still filters, but
-// records no expiry.
-func (t *ProtectionSkipTracker) Skip(log logr.Logger, candidate *kueue.Workload, rule *config.PreemptionProtectionPolicy, reason string, now time.Time) bool {
-	if !WithinProtectionWindow(candidate, rule, now) {
+// (keeping the earliest across calls). Callers must evaluate all other
+// preemption criteria first, so that a skip means protection was the sole
+// blocker. A nil tracker still filters, but records no expiry.
+func (t *ProtectionSkipTracker) Skip(candidate *kueue.Workload, rule *config.PreemptionProtectionPolicy, now time.Time) bool {
+	if !withinProtectionWindow(candidate, rule, now) {
 		return false
 	}
 	expiry := ProtectionExpiry(candidate, rule)
 	if t != nil && (t.earliestExpiry.IsZero() || expiry.Before(t.earliestExpiry)) {
 		t.earliestExpiry = expiry
 	}
-	log.V(4).Info("Skipping preemption candidate within its protection window",
-		"workload", klog.KObj(candidate),
-		"reason", reason,
-		"remainingProtection", expiry.Sub(now))
 	return true
 }
